+++ app/services/delete_environment_variable.rb
@@ -1,6 +1,4 @@
 class DeleteEnvironmentVariable
-  include AwsClientWrapper
-
   attr_accessor :infrastructure
 
   def initialize(infrastructure:)
@@ -10,11 +8,7 @@ class DeleteEnvironmentVariable
   def call(environment_variable:)
     full_name = "#{infrastructure.identifier}/#{environment_variable.full_aws_name}"
 
-    begin
-      aws_ssm_client.delete_parameter(name: full_name)
-      Result.new(true)
-    rescue Aws::SSM::Errors::ParameterNotFound => error
-      Result.new(false, error, "Parameter was not found")
-    end
+    result = DeleteAwsParameter.new(infrastructure: infrastructure).call(path: full_name)
+    result
   end
 end